chore: made any error in read key tasks fatal.
If one of the hosts becomes unavailable for any reason and its key cannot be read,
this will result in the removal of the given instance from the configuration of all other hosts.
If any error is critical and stops the execution of the playbook,
this will ensure the safety of rolling out a new configuration in case of an accidental network flap.